fix: remove circular dependencies form keychain package
