+++ packages/keychain/src/index.ts
@@ -9,6 +9,7 @@ export { decrypt } from './encryption/decrypt';
 export { encrypt } from './encryption/encrypt';
 export * from './profiles';
 export * from './identity';
+export { Profile } from './common';
 
 export default {
   Wallet,
